event_misa/requirements.txt
asgiref==3.6.0
Django==4.1.7
djangorestframework==3.14.0
mssql-django==1.2
mysqlclient==2.1.1
pyodbc==4.0.35
pytz==2022.7.1
sqlparse==0.4.3
